Add support for multi-threaded compilation
